fixing tests/retaining coverage - taking out preview test because it only works with both packages loaded - will discuss

+++ test/gen-zigbee-9.test.js
@@ -90,7 +90,7 @@ test(
     expect(errors.length).toBe(0)
     expect(warnings.length).toBe(0)
   },
-  testUtil.timeout.short()
+  testUtil.timeout.long()
 )
 
 // We here generate each one separatelly in a separate test case.
